Ensure that our systemd units are more tightly bound to the Docker service
BindsTo, as described in systemd docs¹, more accurately describes the
tightly coupled relationship of these units and Docker.

This configuration is still a little odd since the docker client used to
start/stop the containers isn't actually the backend container process
we care about, but it works fairly well in practice and is simple to
deploy.

¹ https://www.freedesktop.org/software/systemd/man/systemd.unit.html#BindsTo=