Start Metabase and Lab Labels services at boot
This was the intended behaviour, but it was thwarted by using an
incorrect, non-existent target unit name (local.target).  Whoops!

Use the standard, built-in, default target instead!

Automatically starting services after boot is important since the
backoffice will now, if necessary, automatically reboot in the middle of
the night to apply security patches.