refresh-database: Automatically create the workdir if explicitly given and it doesn't exist
My pattern is to specify a workdir in temporary directory that's
regularly cleaned out.  This is nice as it means the data is cached
locally to speed up refresh times but doesn't linger indefinitely.  When
the workdir is periodically removed for being stale, the same
refresh-database command will now still work.