id3c-restricted: A second production environment for restricted data
This environment will be used to receive data with more access
restrictions on it.  Unlike our primary production environment, the
database in the restricted environment will not be accessible to most of
the Study staff.

I'd prefer to have more granular access controls in ID3C so we could use
the same environment, but since that requires additional development,
this is a quicker solution at the moment to start accepting data into
receiving.