crontabs: update schedule for SFS REDCap projects
These projects are ingested every hour, interleaving the projects so
each gets ingested every 6 hours.

We are keeping these around in case there are any data corrections made.
The only active REDCap project is the uw-retros. The etl for this
project really only needs to run once a day after the DET generation
at 4 A.M.