Generate lab result report PDFs for SCAN
This will generate new reports every time it runs, instead of only
generating reports for new results or updated results.  This is much
simpler than doing a diff and incremental generation, and it should be
tenable for the near future, and maybe quite a while longer.  I plan to
keep an eye on it to see at what point we may want to switch to
incremental generation.  It's worth noting too that data transfer into
S3 is free, so it won't incur those extra costs for Lab Med.

Barcodes in the results file are now uppercased so the report filenames
are correct.  SecureLink normalizes to uppercase, and Nik's ingest
process for SCAN results was doing this anyway on the results side.