Disable parallel generation of SCAN result PDFs for now
There is some bug in the S3-handling with parallel processes that
doesn't occur for me locally but does in production.  This is a little
insane as both are run in the exact same Docker image.  Not sure if this
is an issue in our code, s3fs, botocore, or even some network-level
difference because production is inside AWS and my laptop is not.