SCAN: Quick-and-dirty stop-gap resolution for PDF generation time issues
The current PDF generation time is ~90 minutes.  This is an issue
because jobs launched every 60 minutes thus pile up and compound the
load on the system.

Moves the jobs to every other hour instead of every hour so jobs don't
overlap at the current runtime.  Then, adds a lockfile so that even if
generation time increases again, the jobs won't overlap.  (We'll see an
error from cron if this happens because flock will exit with an error.)

Better solutions include:

  1. Only generating what we need to, thus making generation much
     faster, albeit at the expense of having to (correctly!) keep track
     of what needs updating.

  2. Generating on a node with more CPUs so we can do more in parallel.