SCAN: Combine data and PDF generation into one cronjob
This allows us to most easily ensure the serial sequencing of each step
so that the results generation is always complete before we generate
PDFs from its output.  We want to keep the data and PDFs in sync,
although we only control part of that process since we're not directly
writing the final JSON files that SecureLink uses (at UW Lab Med's
request).