Rename the `import` script to `upload`
The REDCap API refers to the process of POSTing to their endpoint as
either "importing" or "exporting" a data dictionary. In my opinion, that
that vocabularly makes sense from REDCap's perspective. From our
perpsective as users, I believe the terms "uploading" and "downloading"
a data dictionary to/from REDCap are more appropriate here.

Rename the "import-" prefix of the REDCap data dictionary processing
script to be named "upload-" accordingly.