Add default functionality to argparse
Right now there are rather cryptic error messages when invoking the
import script when certain arguments aren't provided. Fix this by either
providing a default argument (in the case of <field_names.csv>), or
declare that certain arguments are required.

Note that according to the python argparse docs¹, using an `nargs` of
`+` should throw an error if no argument is provided. But, because this
was not happening as expected, I added a `default=True` argument to the
parser for `--html-attriutes`.

¹: https://docs.python.org/3/library/argparse.html#nargs