Remove batch script for data dictionaries
We're modifying already-translated REDCap data dictionaries in-place.
This means that the data dictionary upload is a one-time operation, and
that the timing will differ for each project depending on when finalized
translations are available.

Therefore, it doesn't make sense to actually have this batch script
available in the repo. If used, it should only be invoked once per
REDCap project.