Diff SCAN results files without headers
When diffing, we don't want to duplicate the results files' headers if
the header changed.

Use comm instead of grep, because grep will interpret lines in the file
as patterns instead of fixed strings.

Tom mentions that there's a `-F/--fixed-strings` option in GNU grep but
that it caused him tons of issues in BSD grep when combining `-v`, `-f`,
and `-F`.

To make our lives easier, use comm for diffing instead.