Set up crontab for SCAN Amazon addresses
According to Misja, outgoing Amazon Care addresses should be finalized
for enrollments per day by 6 AM.

There does not seem to be a way to set the role ARN as an environment
variable. We need to set up a profile within the `~/.aws/config` file
that retrieves the source credentials from environment variables.