Disable Amazon Care cronjob
Our upload permissions were revoked shortly after the SCAN pause and
this job has been failing since then.  With the SCAN relaunch, we'll be
integrating with a different group, Amazon Flex.  I'm leaving the rest
of the code in place for now since I suspect we'll shortly re-name and
re-use it.