crontabs: Use "chronic fatigue" or "fatigue --quiet" for all frequent jobs
This avoids flooding our notifications with repeated error messages from
jobs that run very often.  For jobs using chronic already, just chain in
fatigue too.  For jobs which don't use chronic because we want to see
warnings not just errors, use the more selectively-suppressive "fatigue
--quiet".

fatigue is a program I wrote today for this purpose, which I'm tracking
in source control at <https://github.com/tsibley/fatigue> but is just
vendored directly here so it can be tweaked as necessary.