scan-return-of-results: Use NDJSON for REDCap records instead of CSV
This format is less compact, but it makes the data concatenation robust
to differences in field order or even differences in the fields the
exist between REDCap projects.  Both properties are necessary to add the
SCAN kiosks/in-person enrollments project.

Adds an implicit dependency on jq to unwrap REDCap's JSON array response
into NDJSON.