scan-return-of-results: Add REDCap project for kiosk/in-person enrollments
This addition requires quite a bit of new and nuanced barcode handling.
Even though some barcode field names overlap with other projects, the
projects for mailed-to-home kits record barcodes pretty differently than
this new project.

Correctly handling the barcodes is crucial, since it is what links an
individual person to a result, and we don't want to return the wrong
result.  This new code continues to err on the side of not returning any
result instead of possibly returning the wrong result.