scan-return-of-results: Process new results every 30 minutes
The original bumping of this from every hour → every two hours¹ was due
to our generation process taking a _long_ time because of how it worked.
Now that we have a better, faster generation process², we can run more
often which mean people have to wait less time to see their results.
This delay is most visible to positives, since they get a call before
their result is released to the portal.  Worst case, they had to wait
two hours from the call to the PDF.  Now it's 30 minutes.

Unfortunately we don't have run stats available, so I don't know how
long current jobs are taking.  But flock will exit with an error if it
runs into a previous job that's taking longer than 30 minutes, so we can
adjust from there.

¹ "SCAN: Quick-and-dirty stop-gap resolution for PDF generation time issues" (21b87cf)
² "SCAN RoR: Generate PDFs for new results only" (4e84e7e)