scan-return-of-results/transform: set null swab_type to "pending" status
This is done to prevent the return of results from being blocked by the
lack of swab type for never tested samples. The timing of a sample
getting marked as `never-tested` in REDCap and added to the
incident report is out of our control. This may lead to a block of
results since we cannot generate a PDF report without a known swab type.

A separate query is set up on Metabase to check hourly for null swab
types so that we know when we need to manually intervene.