Rebrand SCAN RoR as plain RoR
We are now moving the branding for SCAN under the SFS umbrella. Call the
RoR workflow, simply, "return of results" or "SFS RoR", and drop the
SCAN language.

The new view for results is `shipping.return_results_v3`. There are
currently no results outside of the SCAN IRB study to return.