return-of-results/transform: include UW-reopening REDCap data
Add a new argument for UW-reopening REDCap data and parse data to the
same specifications for the UW Lab Med's return of results portal.

Only mark postives & inconclusives as "pending" if `contacted` is not
"yes" for SCAN results. UW-reopening results can be pushed to the
portal without previous contact from EH&S.