return-of-results: fix SFS longitudinal config file path
Switch the order of `resolve` and `parent` to get the correct file path
for the `config_file`.

From @tsibley:
General rule of thumb is to always `resolve()` first before doing parent
traversals. The order of `parent` vs. `resolve` matters if the path has
symlinks or is relative. In this case, `__file__` is relative because
the cwd is its containing dir.