Refresh Switchboard every 30 minutes
The barcodes that the Switchboard pulls from REDCap and provides
via querying are entered into REDCap when we ship out a kit,
when the participant registers the kit, or when a sample is
collected at a kiosk.

We don't need to pull REDCap records in near real time for Switchboard.
This commit changes the frequency of this job from every 10 minutes to
every 30 minutes. I was going to go for every hour but thought that
it's possible for a sample to get back to the lab from a kiosk on
campus faster than that.