Run fatigue on RoR generation and EHS cronjobs
Now that fatigue is modified to support repeat warnings
and errors, use fatigue --quiet on the return of results
file generation and HCT EHS file transfer processes. This
will notify only once for the same message and make our
alerts channel much more readable when its not overflowing
with the same message.