refresh-database: Exclude comments on extensions
Remove comments on extensions from the TOC when refreshing a target
database with a copy of the production database.

@tsibley pointed out that comments on extensions require a true
superuser, but RDS doesn't offer us that. Therefore, filter out comments
on extensions while preserving other comments (instead of using the
nuclear `--no-comment` option available in Postgres 11).

Note that filtering the TOC avoids the need to detect these errors from
being printed to stdout or stderr, so delete the related regex code
used to suppress errors.