refresh-database: Filter TOC to prevent known errors restoring public.spatial_ref_sys
Filter `TABLE DATA public spatial_ref_sys rdsadmin` from the TOC when
refreshing a target database with a copy of the production database, as
it is owned by rdsadmin, which does not exist locally.

Delete the redundant regex code used to suppress error messages from
stderr and stdout.

Invoke `grep` with the `--extended-regexp` option (`-E`) to interpret
the `|` as an alternation.