Remove unused Amazon Care fulfillment cronjob
Disabled back in June with de7630a, with the expectation we'd re-enable
it again soon.  That never happened.  Time to relegate the unused code
to history.  I uncovered this while auditing all REDCap API credential
usages.

The crontabs Makefile now knows how to delete previously-installed
copies of old crontabs, so that this isn't a manual step we can forget
on deploy.