Use flock for refresh materialized view jobs
To prevent multiple instances of the refresh materialized view jobs
from running at the same time, use flock. Use `--nonblock` so that the
job will fail instead of waiting for the running job to finish. These
jobs run frequently, so we can wait until the next scheduled
run instead of queueing up runs. Use `--conflict-exit-code 0` so that
the job reports as success if it fails because it cannot acquire the
lock.