Filter for samples collected under IRB pre_analytical_specimen_collection
Use the `pre_analytical_specimen_collection` to filter the generation of PDFs for
samples collected under the `IRB` context.

** CAUTION: Do not merge this commit into Master without planning to update
the latest file in S3 with the new pre_analytical_specimen_collection
column as a new column added to the right. If we don't do this correctly,
the first diff will contain all results.