Handle AFH REDCap data and a new PDF template
This commit adds a configuration driven process to bring in SFS
REDCap projects that are in the classic mode, mirroring what is done
already for SFS longitudinal projects. New columns are not added
automatically to the final results CSV file, however, because changing
that file needs to be done with care so that the diff generated on the next
run does not include all resutls from day 1.

** CAUTION: Do not merge this commit into Master without planning to update
the latest file in S3 with the new ordering_provider column as
a new column added to the right. If we don't do this correctly, the first
diff will contain all results.