wa-doh-linelists: Use CRLFs
An employee of WA DoH recently asked us to switch back to the "old"
linelist format of using carriage returns in addition to line feeds at
the end of each line in our submitted CSVs. Their system apparently
requires the presence of carriage returns in order to detect the end of
a record. Upon investigating the old format of CSVs with Chris's help,
we indeed see "\r\n" characters between each row in the CSV. Hardcode in
our linelist `transform` script the CRLF that WA DoH expects.