wa-doh-linelists: Drop shelter_name
`shelter_name` is a column that is redundant with `site_name`. The
specifications (under
`./bin/wa-doh-linelists/docs/SFS_DOH_Linelist_Data_Formats.xlsx`) were
outdated, and we do not actually want to submit the `shelter_name`
column to WA DoH. They recently emailed us asking to drop this, as it
was absent in the "old" linelist format (before the `wa-doh-linelists`
pipeline creation).