Use `ignore_index = True` when appending DataFrames
Without this commit, when two REDCap projects contained the same
barcode, the transform code would fail with
`ValueError: cannot reindex from a duplicate axis`
This is because the final DataFrame had repeated index values;
the indexes start at 1 for each project.
Using 'ignore_index = True' with the append statments causes a new
index to be created and incremented on the final appended DataFrame.