Remove bad assumption and combine append commands
My original statement about appending DataFrames that have different
columns was incorrect. The values in the disconcordant columns will be
`NA` for a string column (and `NAN` for other types). There's no need
to replace data after the append. In fact, the `replace` line was
causing this warning:
`FutureWarning: elementwise comparison failed; returning scalar instead,
but in the future will perform elementwise comparison`

Also, the .append() method can take a list of DataFrames to append. It's
more efficient to append the two in the same command.