RoR: Fix bugs for classic project parsing
Don't treat the "contacted" field as required. We do want to pull
records where this doesn't have a value. We will compare its value
to see if we should call the status 'Pending'.

When checking for required fields, first check if there are any required
before doing the `not all` check. If required_fields is an empty list
the `not all` check will return False.