Temporarily disable childcare-2021 etl cron job
It looks like the field illness_q_date was removed from between the pilot and
2021 childcare projects. We didn't catch this during testing, and there may be
more fields that need to be updated.

This ETL was a nice-to-have but not specifically required. Let's temporarily
disable this cron job for now until we have time to debug and diagnose effort
required to fix.