Reduce redcap batch size for UW Reopening ETL
We are still pulling more data than redcap's api can handle at once.
This is showing up as 500 errors from REDCAP api when we try to fetch
records from this project.

Reducing this again, from 150 to 100 records in a batch.
There is a limit to how much this can be reduced. Having team discussions
about alternate strategies.