Handle empty REDCap recordsets in config driven RoR
The code needs to handle the case where the config driven
classic or longitudinal REDCap project fetch returns no records.
When processing the output of the fetch, check if the DataFrame
created from loading the data has rows, and if not, return None
instead of trying to transform that DataFrame.