Process REDCap data in batches for the EH&S Transfer
Previously, the entire output file from the REDCap dump was loaded
at once. This was causing out of memory errors, as the working memory
grew past 9 GB. In this commit, the REDCap file is read in batches and
the content of each batch is processed completely as one batch. Peak
memory consumption as configured is around 160 MB.

The original `transform` script took about 55 seconds to process today's
data. This new version took slighly longer at 70 seconds.

I've tested by comparing the output generated by the master version of
transform against this new version. There are no differences in the output.