id3c-production: Increase uWSGI worker procs from 1 → 4
One process can get stuck too easily servicing a longer-running request.
We saw this yesterday when a Power BI client was repeatedly hitting the
API.  With 2 threads per proc, this should allow 8 concurrent requests
without queuing.  This is small, but as long as our requests remain
quick to service, it should be a fine.