Add Snohomish Schools REDCap fields CSV file
For the Snohomish School District Testing
project, add a CSV of the REDCap fields that should get the language
tag added. I made this list by looking at the data dictionary for the
Russian version of the project and identifying variables for which the
'section_header', 'field_label', 'field_note', or
`Choices, Calculations, OR Slider Labels` column contained
Russian text.