dev/healthbot_azure_sql_database: use varchar(max)
Note the changes to use varchar(max) instead of varchar(400) for text fields of varying length in 
the healthbot sql server database schema. This was motivated by a value for UtmAudience being 
too long for the field, and therefore blocking the pipeline that ingests data from Azure Data 
Storage into the database.