Extend the slow Metabase query killer to the metabase-privileged user
This second database user is configured for the "with hCoV-19" data
source in Metabase, as it is granted different permissions in the
database.

It will go away in the future as this split is no longer needed, but
it's still here now and is the source of many runaway queries.