Automate linelist generation and upload to S3
Use cron to generate daily linelists and upload to S3. Study staff can
then download the file from S3 to distribute to partner agencies,
instead of having to generate it locally.

This lets us control the environment for linelist generation and
simplifies the process for study staff. Eventually, it'd be great to
automatically upload the linelists directly to the partner agencies.