Exclude queries run by Metabase for sync/scan introspection from the slow-query killer
Otherwise some of these routine queries are not completing for very
field/value enumerations of very large and slow views.  We will
additionally reduce their cadence so they don't run as often.

Queries that are run on behalf of a user have an extended remark comment
at the beginning of the query indicating as much.¹  Positively select
for these to allow any non-user driven queries through.

¹ https://github.com/metabase/metabase/blob/b8c81d0277c509f37ade13a764eef5dc42a06fb6/src/metabase/query_processor/util.clj#L26-L37