Publish weekly specimen counts every day
This data file will power a chart of specimen counts on seattleflu.org.

If we wanted to ensure that the input rows to the aggregated exports for
prevalence and specimens remain identical, I think the best
forward-looking approach would be to export a single, non-aggregated
dataset of:

  (anonymous sample id¹, collection week, site type, pathogen, present)

rows and then use that to drive both charts.  Aggregation could happen
on the fly during charting, or be precomputed (which would be much
speedier for clients to download/display).  However, there's a little
more to consider there from a data publication/privacy perspective² and
so I punted on Just Doing It.

¹ Maybe warehouse.sample.identifier, but not certain. Maybe just a
  serial integer that's not stable across export runs?

² One example: there might be instances of a small n for a given
  (collection week, site type).