Reduce det-limit for UW Reopening ETL to 500 records
Due to other issues, we fell behind in our typical processing of UW
re-opening/ HCT records. As such, we were processing closer to the
prior det limit of 1000 dets at a time, and were seeing OOM on the
host trying to process that many records at once. Typically, this
job had been running far below the set limit of 1000. Let's reduce
the job to process a max of 500 redcap DETs at a time.