Add --upload-if-empty to always upload linelists
Add a new --upload-if-empty argument to allow user to specify a
list of sites we always want to upload the linelist to, even if
it's empty.

The argument can be specified multiple times, and can be a protocol
or destination. If it's a protocol, then any url (or path) using that
protocol will always be uploaded to; if a destination, only that one
site/url/path will be affected.