Add jupyter notebook used to generate Apple JSONs
We agreed to do a one-time upload of symptom data and test results
after data collection stopped for the Apple project. This jupyter
notebook was used to generate the final JSONs uploaded to
Apple's S3 bucket.

This depends on upstream data from a REDCap report and a Metabase query.
Details are in the jupyter notebook comments.