etc/uw-retrospective-manifest: add aliquot sheet 15
Adds aliquot sheet 15 to the list of worksheets for ingestion. Aliquot sheet 15
is the first and likely only worksheet that will be generated directly
from LIMS as an interim solution until we have an API developed accepting data in
real time.

This YAML file also switches to the new simplified format where multiple worksheets
sharing the same format can be entered as a list of values.