Add SSH creds to linelist crontab
Add SSH username & private key file to wa-doh-linelist crontab for
automated file upload to the state Department of Health. It was
previously manually entered for testing; now that we've got it
working, it can go in version control.

The full path to the key is specified because cron failed to
expand $HOME, causing errors.